// src/app/utils/enums.ts
export enum StatusValidacion {
  Init = 0,
  Success = 1,
  Error = 2,
  Warning = 3,
  Loading = 9
}

export enum CatPlantillas {
  NoticiaCriminal = 1,
  CarpetaInvestigacion = 2,
  ActosInvestigacion = 3,
  Delitos = 4,
  Aseguramientos = 5,
  VictimasDelito = 6,
  ImputadoDelito = 7,
  VictimaImputado = 8,
  Determinacion = 9,
  Proceso = 10,
  MandamientosJudiciales = 11,
  InvestigacionComplementaria = 12,
  MedidasCautelares = 13,
  EtapaIntermedia = 14,
  MASC = 15,
  Sobreseimiento = 16,
  SuspensionCondicional = 17,
  Sentencia = 18
}